Fixes reset profile dialog.
Hides the footer that offers to send settings to Brave.

Fixes brave/brave-browser#18077
