+++ browser/resources/settings/BUILD.gn
@@ -118,6 +118,7 @@ preprocess_if_expr("preprocess") {
     "brave_overrides/site_settings_page.js",
     "brave_overrides/sync_account_control.js",
     "brave_overrides/sync_controls.js",
+    "brave_reset_page/brave_reset_profile_dialog_behavior.js",
     "brave_routes.js",
     "brave_sync_page/brave_sync_browser_proxy.js",
   ]
