+++ browser/resources/settings/brave_overrides/index.js
@@ -28,6 +28,7 @@ import './people_page.js'
 import './personalization_options.js'
 import './printing_page.js'
 import './privacy_page.js'
+import './reset_profile_dialog.js'
 import './safety_check_passwords_child.js'
 import './search_engines_section.js'
 import './security_page.js'
