+++ browser/resources/settings/brave_overrides/reset_profile_dialog.js
@@ -3,24 +3,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // you can obtain one at http://mozilla.org/MPL/2.0/.
 
+import {BraveResetProfileDialogBehavior} from '../brave_reset_page/brave_reset_profile_dialog_behavior.js'
 import {RegisterPolymerComponentBehaviors} from 'chrome://brave-resources/polymer_overriding.js'
 
 RegisterPolymerComponentBehaviors({
   'settings-reset-profile-dialog': [
-    ready: function() {
-      this.fixSendSettingsCheckbox();
-    },
-
-    /**
-     * Unchecks, disables, and hides sendSettings checkbox.
-     * Also, hide the entire footer.
-     * @private
-     */
-    fixSendSettingsCheckbox: function() {
-      this.$.sendSettings.checked = false;
-      this.$.sendSettings.hidden = true;
-      this.$.sendSettings.disabled = true;
-      this.$.sendSettings.parentNode.hidden = true;
-    }
+    BraveResetProfileDialogBehavior
   ]
 })
