+++ browser/resources/settings/brave_reset_page/brave_reset_profile_dialog_behavior.js
@@ -1,0 +1,26 @@
+/* Copyright (c) 2021 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+const BraveResetProfileDialogBehaviorImpl = {
+  ready: function() {
+    this.fixSendSettingsCheckbox();
+  },
+
+  /**
+   * Unchecks, disables, and hides sendSettings checkbox.
+   * Also, hide the entire footer.
+   * @private
+   */
+  fixSendSettingsCheckbox: function() {
+    this.$.sendSettings.checked = false;
+    this.$.sendSettings.hidden = true;
+    this.$.sendSettings.disabled = true;
+    this.$.sendSettings.parentNode.hidden = true;
+  }
+};
+
+export const BraveResetProfileDialogBehavior = [
+  BraveResetProfileDialogBehaviorImpl
+]
